add storing RefreshTokenInfo in SubjectCredentialStore.Entry
